refactor + fix filters
