+++ messages/pt-BR.json
@@ -578,6 +578,8 @@
 	"share_importMods": "{count} mods",
 	"share_importMissing": "{count} mod(s) não encontrado(s)",
 	"share_importFrom": "Perfil: {name}",
+	"share_importSuccess": "Perfil \"{name}\" importado!",
+	"share_importOverwriteSuccess": "Mods importados no perfil atual!",
 	"share_profileName": "Nome do perfil",
 	"prefs_syncComingSoon": "A sincronização ainda não está disponível. Em breve!",
 	"profiles_uploadingIcon": "Enviando foto de perfil...",
