+++ src/lib/components/ui/Checkbox.svelte
@@ -25,7 +25,13 @@
 >
 	{#if checked}
 		<svg viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-			<path d="M2.5 6L5 8.5L9.5 3.5" stroke="currentColor" stroke-width="1.8" stroke-linecap="round" stroke-linejoin="round" />
+			<path
+				d="M2.5 6L5 8.5L9.5 3.5"
+				stroke="currentColor"
+				stroke-width="1.8"
+				stroke-linecap="round"
+				stroke-linejoin="round"
+			/>
 		</svg>
 	{/if}
 </button>
