+++ src/lib/utils/file.ts
@@ -1,0 +1,11 @@
+export function fileToBase64(file: File): Promise<string> {
+	return new Promise((resolve, reject) => {
+		const reader = new FileReader();
+		reader.readAsDataURL(file);
+		reader.onload = () => {
+			const result = reader.result as string;
+			resolve(result.split(',')[1]);
+		};
+		reader.onerror = (error) => reject(error);
+	});
+}
