+++ src/lib/utils/select.ts
@@ -1,0 +1,6 @@
+export function selectItems(
+	items: string[],
+	getLabel: (item: string) => string = (value) => value as string
+) {
+	return items.map((item) => ({ value: item, label: getLabel(item) }));
+}
