+++ src/routes/+layout.svelte
@@ -72,6 +72,13 @@
 				lang = prefs.language;
 			}
 
+			// Fallback to base locale if stored language was removed
+			if (!locales.includes(lang as Locale)) {
+				lang = 'en';
+				prefs.language = lang;
+				await api.prefs.set(prefs);
+			}
+
 			if (lang !== getLocale()) {
 				updateAppLanguage(lang);
 			}
@@ -101,17 +108,29 @@
 	onkeydown={(evt) => {
 		const k = evt.key.toLowerCase();
 		// Block F12 (devtools)
-		if (k === 'f12') { evt.preventDefault(); return; }
+		if (k === 'f12') {
+			evt.preventDefault();
+			return;
+		}
 		// Block F5 (refresh via F5)
-		if (k === 'f5') { evt.preventDefault(); return; }
+		if (k === 'f5') {
+			evt.preventDefault();
+			return;
+		}
 		// Block Ctrl+shortcuts except Ctrl+R (reload), Ctrl+C/V/X/A/Z (standard editing)
 		if (evt.ctrlKey && !evt.shiftKey && !evt.altKey) {
 			const allowed = ['r', 'c', 'v', 'x', 'a', 'z'];
-			if (!allowed.includes(k)) { evt.preventDefault(); return; }
+			if (!allowed.includes(k)) {
+				evt.preventDefault();
+				return;
+			}
 		}
 		// Block Ctrl+Shift+I/J/C (devtools variants)
 		if (evt.ctrlKey && evt.shiftKey) {
-			if (['i', 'j', 'c'].includes(k)) { evt.preventDefault(); return; }
+			if (['i', 'j', 'c'].includes(k)) {
+				evt.preventDefault();
+				return;
+			}
 		}
 	}}
 />
