+++ src/routes/prefs/+page.svelte
@@ -324,7 +324,7 @@
 		padding: var(--space-xl) var(--space-xl) 0;
 	}
 
-	.appearance_margin{
+	.appearance_margin {
 		margin-top: var(--space-xl);
 	}
